src/requirements.txt
asyncpg==0.22.0
fastapi==0.63.0
gunicorn==20.0.4
newspaper3k==0.2.8
requests==2.25.1
tortoise-orm==0.16.21
uvicorn==0.13.4
python-multipart==0.0.5
python-jose[cryptography]==3.2.0
passlib[bcrypt]==1.7.4
